fixed flags not loading

+++ app/src/main/java/com/shubh/covid19tracker/CountryInfo.kt
@@ -4,7 +4,7 @@ data class CountryInfo(
 	val flag: String? = null,
 	val id: Int? = null,
 	val iso2: String? = null,
-	val lat: Int? = null,
+	val lat: Float? = null,
 	val jsonMemberLong: Int? = null,
 	val iso3: String? = null
 )
